+++ CHANGELOG.md
@@ -9,12 +9,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
-- [#452] All packages now use the `"exports"` field in their `package.json` for better ESM compatibility.
 - [#455] A UMD build is no longer provided. See [new instructions for buildless environments using ESM](https://react-querybuilder.js.org/docs/buildless)).
 - [#431] Major `ValueEditor` update--including the `ValueEditor`s in the compatibility packages--for "between"/"notBetween" operators. When the `operator` for a rule is "between" or "notBetween", two inputs will be displayed. Each will have the class "rule-value-list-item". They will manage the `value` as a comma-separated list unless `listsAsArrays` is `true`, in which case a proper array will be used.
 - [#431] The default border radius on rule groups and branch lines (SCSS variable `$rqb-border-radius`) is now `0.25rem` (previously `4px`). Visually, this should be the same for most users since `16px` is the default `font-size` on most browsers, and $16 \times 0.25 = 4$.
 - [#431] Utility function `c` has been removed. Use a package like [`clsx`](https://www.npmjs.com/package/clsx) (what RQB uses internally) instead.
 - [#431] Bulma compatibility components no longer specify the `is-small` class, so they will be rendered at their default size.
+
+### Fixed
+
+- [#452] All packages now use the `"exports"` field in their `package.json` for better ESM compatibility.
 
 ### Added